move controller specs to core engine.
