+++ cdx_core/Gemfile
@@ -33,6 +33,7 @@ gem 'cdx-api-elasticsearch', path: '../'
 gem 'cdx-sync-server', git: 'https://github.com/instedd/cdx-sync-server.git', branch: 'master'
 gem 'geojson_import', git: 'https://github.com/instedd/geojson_import', branch: 'master'
 gem 'location_service', git: 'https://github.com/instedd/ruby-location_service.git', branch: 'master'
+gem 'view_components', git: 'https://github.com/manastech/rails-view_components.git', branch: 'master'
 gem 'guid'
 gem 'oj'
 gem 'encryptor'
