+++ cdx_core/spec/controllers/alert_groups_controller_spec.rb
