+++ cdx_core/spec/controllers/device_commands_controller_spec.rb
