+++ cdx_core/spec/controllers/device_logs_controller_spec.rb
