+++ cdx_core/spec/controllers/device_models_controller_spec.rb
