+++ cdx_core/spec/controllers/devices_controller_spec.rb
