+++ cdx_core/spec/controllers/encounters_controller_spec.rb
