+++ cdx_core/spec/controllers/ftp_settings_controller_spec.rb
