+++ cdx_core/spec/controllers/microscopy_results_controller_spec.rb
