+++ cdx_core/spec/controllers/patient_logs_controller_spec.rb
