+++ cdx_core/spec/controllers/patients_controller_spec.rb
