+++ cdx_core/spec/controllers/registrations_controller_spec.rb
