+++ cdx_core/spec/controllers/sample_identifiers_controller_spec.rb
